Rename cartesian_speed.* -> limit_cartesian_speed.*
