+++ moveit_core/trajectory_processing/CMakeLists.txt
@@ -5,7 +5,7 @@ add_library(${MOVEIT_LIB_NAME}
   src/iterative_spline_parameterization.cpp
   src/trajectory_tools.cpp
   src/time_optimal_trajectory_generation.cpp
-  src/cartesian_speed.cpp
+  src/limit_cartesian_speed.cpp
 )
 set_target_properties(${MOVEIT_LIB_NAME} PROPERTIES VERSION "${${PROJECT_NAME}_VERSION}")
 
@@ -24,6 +24,6 @@ if(CATKIN_ENABLE_TESTING)
   target_link_libraries(test_time_parameterization moveit_test_utils ${catkin_LIBRARIES} ${urdfdom_LIBRARIES} ${urdfdom_headers_LIBRARIES} ${MOVEIT_LIB_NAME})
   catkin_add_gtest(test_time_optimal_trajectory_generation test/test_time_optimal_trajectory_generation.cpp)
   target_link_libraries(test_time_optimal_trajectory_generation moveit_test_utils ${catkin_LIBRARIES} ${console_bridge_LIBRARIES} ${MOVEIT_LIB_NAME})
-  catkin_add_gtest(test_cartesian_speed test/test_cartesian_speed.cpp)
-  target_link_libraries(test_cartesian_speed moveit_test_utils ${catkin_LIBRARIES} ${urdfdom_LIBRARIES} ${urdfdom_headers_LIBRARIES} ${MOVEIT_LIB_NAME})
+  catkin_add_gtest(test_limit_cartesian_speed test/test_limit_cartesian_speed.cpp)
+  target_link_libraries(test_limit_cartesian_speed moveit_test_utils ${catkin_LIBRARIES} ${urdfdom_LIBRARIES} ${urdfdom_headers_LIBRARIES} ${MOVEIT_LIB_NAME})
 endif()
